refactor to mypy clean
