+++ edg/abstract_parts/Antenna.py
@@ -1,5 +1,7 @@
+from typing import Any
+from typing_extensions import override
+
 from ..electronics_model import *
-from .Categories import *
 from .PartsTable import PartsTableColumn, PartsTableRow
 from .PartsTablePart import PartsTableSelector
