+++ edg/abstract_parts/Crystal.py
@@ -1,8 +1,10 @@
+from typing import Any
+from typing_extensions import override
+
 from ..electronics_model import *
 from .PartsTable import PartsTableColumn, PartsTableRow
 from .PartsTablePart import PartsTableSelector
-from .AbstractCapacitor import Capacitor
-from .Categories import *
+from .Capacitor import Capacitor
 from .StandardFootprint import StandardFootprint, HasStandardFootprint
