+++ edg/abstract_parts/Fuse.py
@@ -1,9 +1,9 @@
-from typing import Optional
+from typing import Any, Optional
+from typing_extensions import override
 
 from deprecated import deprecated
 
 from ..electronics_model import *
-from .Categories import *
 from .PartsTable import PartsTableColumn, PartsTableRow
 from .PartsTablePart import PartsTableSelector
 from .StandardFootprint import StandardFootprint, HasStandardFootprint
