+++ edg/abstract_parts/IoControllerProgramming.py
@@ -3,7 +3,7 @@ from typing import List
 from typing_extensions import override
 
 from ..electronics_model import *
-from .AbstractDebugHeaders import (
+from .SwdCortexConnectors import (
     SwdCortexTargetConnector,
     SwdCortexTargetConnectorReset,
     SwdCortexTargetConnectorSwo,
