+++ edg/abstract_parts/Jumper.py
@@ -1,7 +1,6 @@
 from typing_extensions import override
 
 from ..electronics_model import *
-from .Categories import DiscreteComponent, TypedJumper
 
 
 @abstract_block
