+++ edg/abstract_parts/Opamp.py
@@ -3,7 +3,6 @@ from typing import Mapping, Tuple, List, NamedTuple
 from typing_extensions import override
 
 from ..electronics_model import *
-from .Categories import Analog, MultipackDevice
 
 
 @abstract_block
