+++ edg/abstract_parts/Oscillator.py
@@ -1,9 +1,9 @@
+from typing import Any
 from typing_extensions import override
 
 from ..electronics_model import *
 from .PartsTable import PartsTableColumn, PartsTableRow
 from .PartsTablePart import PartsTableSelector
-from .Categories import *
 
 
 @abstract_block
