+++ edg/abstract_parts/Resistor.py
@@ -1,8 +1,8 @@
 import re
-from typing import Optional, Mapping, Dict
+from typing import Any, Optional, Mapping, Dict
+from typing_extensions import override
 
 from ..electronics_model import *
-from .Categories import *
 from .ESeriesUtil import ESeriesUtil
 from .PartsTable import PartsTableColumn, PartsTableRow
 from .PartsTablePart import PartsTableSelector
