+++ edg/abstract_parts/ResistorArray.py
@@ -1,6 +1,8 @@
+from typing import Any
+from typing_extensions import override
+
 from ..electronics_model import *
-from .AbstractResistor import Resistor
-from .Categories import *
+from .Resistor import Resistor
 from .PartsTable import PartsTableColumn, PartsTableRow
 from .PartsTablePart import PartsTableSelector
 from .StandardFootprint import StandardFootprint, HasStandardFootprint
