+++ edg/abstract_parts/SpiMemory.py
@@ -1,7 +1,6 @@
 from typing import Any
 
 from ..electronics_model import *
-from .Categories import *
 
 
 @abstract_block
