+++ edg/abstract_parts/Switch.py
@@ -1,9 +1,7 @@
-from typing import Dict, Any
-
+from typing import Any, Dict
 from typing_extensions import override
 
 from ..electronics_model import *
-from .Categories import *
 
 
 @abstract_block
