+++ edg/electronics_interfaces/DummyDevices.py
@@ -1,7 +1,7 @@
 from typing import Dict
+from typing_extensions import override
 
 from ..electronics_model import *
-from edg.abstract_parts.Categories import *
 
 
 class DummyPassive(DummyDevice):
