+++ edg/electronics_interfaces/MergedBlocks.py
@@ -1,9 +1,7 @@
 from typing import Dict
-
 from typing_extensions import override
 
 from ..electronics_model import *
-from edg.abstract_parts.Categories import *
 
 
 class MergedVoltageSource(DummyDevice, GeneratorBlock):
