+++ edg/electronics_model/Categories.py
@@ -2,8 +2,7 @@ from typing import Any
 
 from typing_extensions import override
 
-from ..core import BoolLike, init_in_parent
-from ..electronics_model import Block, abstract_block, InternalBlock
+from ..core import BoolLike, Block, abstract_block, InternalBlock
 
 
 @abstract_block
